Updated: memory_manager.py
I updated the allocation function to include the ability to traverse the list and find empty memory blocks, which were actually nodes, of the requested size or to split an existing one so however much size you wanted would be taken from it and put into the new block, and the other gets that size subtracted from it. This was done using linked lists, loops, and iteration.