Updated: memory_manager.py
I patched some bugs related to the head node being deleted and then referred to, which resulted in ValueErrors, and with referring to certain objects as Booleans when they were not, which caused some errors that happened behind-the-scenes. Along with restructuring the entirety of the module's code. Specifically the allocate() function due to many faulty lines of code.